Implemented basic TensorBoard for cont.learn
Looked through TensorFlow and TensorBoard documentation, and replicated
callback implementation.  The code instantiates a TensorBoard object to
a generic "logs" directory, based on datetime of start (down to seconds)
and logs for each batch/epoch the current loss for the session.  Some
members are added to the CameraStreamingMainWindow for tracking epoch
and loss information, to emulate the keras.model.fit implementation.
Each fit session is considered an epoch with only a single batch. If the
continuous learning button is pressed again to terminate cont.learn,
then that terminates the learning session for the TensorBoard module,
and the callback list is cleared back to None.

Future implementations will need to separate the continuous learning
module from the camera streaming main window module, meaning that this
code will need to be encapsulated at that point (the MainWindow code is
doing too much and needs to be refactored).