66 modify dataset loader to correctly track filepaths (#91)
* WIP: Modifies loaders to return dictionaries

The loaders now pass a dictionary of "image" and "filepath" for each
entry.  The issue now is how to unify all the code to handle this
expectation.  Several test scripts will need to be written to handle the
modification and change without negatively impacting the code. Some
unification needs to be done to ensure that training can handle both
expected benchmark TFDS datasets and the new modification.  Fileorder
needs to also be evaluated to ensure that the file order is indeed
correct.

* Adds split for img/filepaths