64 implement offline latent space cosine similarity analysis of model and dataset (#92)
* BUG: Comment out bbox annotations

Attempting to do analysis on files loaded from recording produced a bug,
where the bbx is attempted to be read for the annotations.  Annotations
are out of format and need to be modified to maintain previous
implementation, but prior implementation was already disabled due to
slow processing time with RaggedTensors.  Therefore, for now, just
comment out the code until I can get to it.

* WIP: Implemented distance metrics without output to file

Was able to implement cosine distance and euclidean distance, and plot
to matplotlib histograms. Ran into an issue with trying to extract
original filenames from data loader.  Due to the complexities of caches
and interleaving from TensorFlow Datasets, the file order that I
believed I was keeping track of is not stable and does not align with
the sample inputs.  Since this is not a training task, it would make
sense to instead of using TFDS, use a simpler module that simply loads
the images and filenames from the JSON. This is a limitation and
short-sight due to the added complexity of using TensorFlow datasets.

* WIP: Adds plot of distances and latent histograms

Still no output to file

* BUG: Comment out bbox annotations

Attempting to do analysis on files loaded from recording produced a bug,
where the bbx is attempted to be read for the annotations.  Annotations
are out of format and need to be modified to maintain previous
implementation, but prior implementation was already disabled due to
slow processing time with RaggedTensors.  Therefore, for now, just
comment out the code until I can get to it.

* WIP: Implemented distance metrics without output to file

Was able to implement cosine distance and euclidean distance, and plot
to matplotlib histograms. Ran into an issue with trying to extract
original filenames from data loader.  Due to the complexities of caches
and interleaving from TensorFlow Datasets, the file order that I
believed I was keeping track of is not stable and does not align with
the sample inputs.  Since this is not a training task, it would make
sense to instead of using TFDS, use a simpler module that simply loads
the images and filenames from the JSON. This is a limitation and
short-sight due to the added complexity of using TensorFlow datasets.

* WIP: Adds plot of distances and latent histograms

Still no output to file

* Outputs plots and distances to file

* Standardize latent space distances

* Adds metrics and standardization flags

* Add standardized transform for embeddings